ats_resume_optimizer/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
nltk==3.8.1
scikit-learn
openai==1.30.1
python-dotenv==1.0.1
docx2txt==0.8
reportlab==4.2.2
